requirements/base.txt
# Basic stuff
django>=1.8.3,<1.9

# Templates
django-compressor==1.5

# Database
psycopg2==2.6
dj-database-url==0.3.0

# Styling
django-flat-theme==0.9.3
django-bootstrap-form==3.2

# Bcrypt passwords
bcrypt==1.1.1

# Kitchen sinks
django-braces==1.8.1
django-extensions==1.5.5

# Data import
django-csvimport==2.4

# Auth
django-allauth==0.23.0
oauthlib==1.0.1
python-openid==2.2.5
requests-oauthlib==0.5.0
